Don't let display cutout safe area be affected by status bar
Otherwise, we will get a big gap at the top edge of display while
playing landscape games on a device with waterfall cutout.

The original logic was to avoid that fullscreen apps will cut halfway
into the status bar. However, for non-waterfall cases, it can be avoided
by setting the height of display cutout the same as the height of status
bar. For the waterfall cases, althrough apps cut halfway into the status
bar, but they don't cut into the icons of status bar. This should be
fine.

This CL also removes ITYPEs from WindowLayout and WindowLayoutTests, and
moves WindowLayoutTests to coretests.

Fix: 264640845
Bug: 234093736
Test: atest WindowLayoutTests
Change-Id: Iaec9bf50b45f8f6e27695a4267cf323ae640c680