Make mutable PendingIntent explicit
Starting from target SDK U, we will block creation of mutable
PendingIntents with implicit Intents because attackers can mutate the
Intent object within and launch altered behavior on behalf of victim
apps. For more details on the vulnerability, see go/pendingintent-rca.

From a quick analysis, we concluded that the PendingIntent here was only
destined to the test app/to the app, so it was made explicit. Reviewers,
please call out if this is not the case.

Bug: 236704164
Bug: 229362273
Test: atest RemoteInputViewTest
Test: atest InflatedSmartRepliesTest
Test: atest HeadsUpManagerTest
Test: atest SmartReplyViewTest
Test: atest NotificationUiAdjustmentTest
Test: atest NotificationInterruptStateProviderImplTest
Change-Id: Icfc653910fac99acfab74bb513d0df5b73183ca7