Fix notification permission for letterboxed activity
The transparent activities letterbox policy was not applied
because notification permission requests are launched from
a process for a uid  different from the one associated with
the first opaque activity. This led to misalignment between
a notification permission dialog and an activity below it.

This CL removes a condition that checked that launchedFromUid
and UID of the first opaque activity are the same.

Fixes: 265116754
Test: Launch an app with a notification permission request dialog
anche check it's cenetered with the letterboxed activity below. Also
`atest WmTests:SizeCompatTests`

Change-Id: I2c50b55fe99e189ce6f4c5efe7a20e58c19d8660