Add explicit handling for an unspecified gwpAsanMode manifest flag.
In future, we'd like to change the default for an unspecified
gwpAsanMode flag. Currently, this implies
mallopt(M_INITIALIZE_GWP_ASAN, { .desire =
DONT_TURN_ON_UNLESS_OVERRIDDEN }) for non-system apps and { .desire =
TURN_ON_WITH_SAMPLING } for system apps.

But, as noted in the TODO, we'd like to change this behaviour. Thus, we
need to plumb through the fact that the manifest flag was left
unspecified.

Bug: 247012630
Test: Add some logging code to libc's android_mallopt, install an app
with an unspecified gwpAsanMode, and make sure that log line is hit.

Change-Id: I45b064c89ceeace0014a19ba1d9e8ac0a2157328