Fix app compat component properties documentation.
Update documentation to correctly indicate that properties are applied on the app level rather than the activity level. Applying properties on the app level maintains consistency with the level on which per-app overrides are applied by device manufacturers. Also, this allows to easily opt-out/in the whole app from per-app overrides.

Fix: 267082980
Test: builds
Change-Id: Icb2b4394b73216c0af762db07dca42796eb4d49a