Update FP failure logic when device is unlocked
* Update strings when fingerprint auth fails
when the device is already unlocked (by face auth or
a TrustAgent). Instead of telling the user fingerprint failed,
instead tell the user why the device is unlocked and how
to enter the device (ie: Unlocked by Face, Swipe up to open)
* If fingerprint gets locked out while face (nonbypass) or a
TrustAgent is keeping the device unlocked - don't allow the user
to use the face or TrustAgent trust to enter the device. Instead,
require primary authentication to enter the device on fingerprint
lockout.

Test: atest KeyguardUpdateMonitorTest KeyguardIndicationControllerTest
Fixes: 265443040
Change-Id: Ibe89cb8703fdf82f4f3f22c556b32fd9fcc3b2c8