Schedule idle timeout when Keyguard state is reset and only if possible
The 4-hour idle timeout is scheduled when the device is locked (by
manually pressing the power button or after the screen timeout). This
usually happends when Keyguard decides to show up.

In the face non-bypass case, Keyguard is already shown and its state
will be reset after locking device. The idle timeout should also be
scheduled in this case.

Additionally, check if any non-strong biometrics are enrolled and
enabled. If not, do not schedule the idle timeout when the device is
locked.

Bug: 244743493, 261813434
Test: On a device with UDFPS, manually
      (1) set up face auth, (2) unlock with face auth and stay on
      Keyguard, (3) lock device, and verify that the alarm for idle
      timeout is scheduled properly by "adb shell dumpsys alarm"
Test: On a device with side FPS, manually
      (1) set up fingerprint (no face auth), (2) lock device, and verify
      that no alarm is schedule for idle timeout, (3) swipe up to show
      bouncer, and verify that the string on bouncer properly shows
      "Enter your PIN" while side FPS can also be used for auth.
      Repeat steps 2-3 after enabling face auth and then deleting it.

Change-Id: I8404927e8155f40496ac1740e4a6b7b38458f1f5