Log instanceId with stylus UIEvents.
For bluetooth styluses, an instance ID represents a single session where a single bluetooth stylus connects and disconnects. This distinguishes between multiple styluses. For USI styluses, the presence and then removal of any USI battery represents a single session/instance ID, and does not distinguish multiple styluses.

For USI stylus notifications, a unique instance ID is generated when a notification is first shown, and is reset when the same notification is cancelled, whether through user dismissal or e.g. connection of a bluetooth stylus, representing the 'lifetime' of the notification.

Bug: 267815315
Test: StylusUsiPowerUiTest, StylusManagerTest
Change-Id: Ie34143328a6a44a0086bec8ceb4f9502e8d4e81f