Move lower idle alarm throttling to rolling quotas
ALLOW_WHILE_IDLE_COMPAT_QUOTA is used for throttling the following
allowed-while-idle alarms:
- Inexact alarms that use a PendingIntent
- Listener alarms
- Exact alarms for pre-S apps
- Alarms from system power allowlisted apps

This quota has stayed at 1/9 minutes for legacy reasons. Fixed delay
throttling is suitable for a small subset of apps that have a sustained
presence and require periodic alarms for a long duration of time. Such
use cases can now move to the setPrioritized API.
Moving this quota to an equivalent rolling window quota of 7/hour since
that would generally be better suited for a large majority of apps as
it allows greater flexibility within the window.

Test: atest FrameworksMockingServicesTests:AlarmManagerServiceTest
Test: atest CtsAlarmManagerTestCases
Test: atest CtsBatterySavingTestCases:BatterySaverAlarmTest

Bug: 209514618
Change-Id: Iad8c3a4f652d1a95bd0deb9b49ecefa6f8bf5b64