Handle repeated uncertainty better
When a primary LTZP reports it is uncertain, the uncertainty timer is
started in the location_time_zone_manager before the uncertainty is
passed on.  This gives the primary LTZP a period to "change its mind"
and also the secondard is started. Once the timer completes without the
uncertainty being contradicted the uncertainty is passed on to the
time_zone_detector. "Certain" reports are always passed on immediately
and cause the timer to be cancelled.

The old logic would repeat this timeout logic even if the
location_time_zone_manager had already completed the uncertainty timeout
and reported it was uncertain.

The new logic will pass on uncertainty immediately if the
location_time_zone_manager is already uncertain. There's no obvious
benefit to starting the timeout again and it looks confusing in the
logs.

Bug: 268162169
Test: atest services/tests/servicestests/src/com/android/server/timezonedetector/location/LocationTimeZoneProviderControllerTest.java
Change-Id: Id596de4067dc41cb1e9d46a920e7610a0eb06933